+++ src/main/java/com/dynatrace/openkit/core/objects/OpenKitImpl.java
@@ -225,6 +225,11 @@ public class OpenKitImpl extends OpenKitComposite implements OpenKit {
     }
 
     @Override
+    public Session createSession() {
+        return createSession(null);
+    }
+
+    @Override
     public void shutdown() {
         if (logger.isDebugEnabled()) {
             logger.debug(getClass().getSimpleName() + " shutdown() - shutdown requested");